XXX: also move the earlier structs into their own package
